b_src/requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0